requirements.txt
python-socketio>=5.0.0
aiohttp>=3.8.0
requests>=2.25.0
